Linux compilation fixes

+++ local/plugin/portaudio/pa_converters.c
@@ -344,10 +344,10 @@ static void Float32_To_Int32(
     {
         /* REVIEW */
 #ifdef PA_USE_C99_LRINTF
-        float scaled = *src * 0x7FFFFFFF;
+        float scaled = *src * (float)0x7FFFFFFF;
         *dest = lrintf(scaled-0.5f);
 #else
-        double scaled = *src * 0x7FFFFFFF;
+        double scaled = *src * (float)0x7FFFFFFF;
         *dest = (PaInt32) scaled;        
 #endif
         
@@ -400,11 +400,11 @@ static void Float32_To_Int32_Clip(
     {
         /* REVIEW */
 #ifdef PA_USE_C99_LRINTF
-        float scaled = *src * 0x7FFFFFFF;
+        float scaled = *src * (float)0x7FFFFFFF;
         PA_CLIP_( scaled, -2147483648.f, 2147483647.f  );
         *dest = lrintf(scaled-0.5f);
 #else
-        double scaled = *src * 0x7FFFFFFF;
+        double scaled = *src * (float)0x7FFFFFFF;
         PA_CLIP_( scaled, -2147483648., 2147483647.  );
         *dest = (PaInt32) scaled;
 #endif
@@ -532,7 +532,7 @@ static void Float32_To_Int24_Clip(
     while( count-- )
     {
         /* convert to 32 bit and drop the low 8 bits */
-        double scaled = *src * 0x7FFFFFFF;
+        double scaled = *src * (float)0x7FFFFFFF;
         PA_CLIP_( scaled, -2147483648., 2147483647.  );
         temp = (PaInt32) scaled;
